Tests: Forbid tests from writing to CWD
Allowing tests writing to the working directory can mask problems. For example, multiple tests running in the same jvm, and using the same relative path, may cause issues if the first test to run leaves data in the directory, and the second test does not remember to cleanup the path before using it.

This change adds security manager rules to disallow tests writing to the working directory. Instead, tests create a temp dir with the existing test framework.

Related to:

* https://github.com/elastic/elasticsearch/commit/a3f078985b1e3675117d134af7fd67ea39302549