Make plugins compile in the eclipse IDE always.
Regardless of the names of other projects in the workspace!
This means elasticsearch-xxx-plugin will always depend on
the .m2/repository/..../elasticsearch.jar even if you happened
to name a project in your workspace 'elasticsearch'. Previously,
it would add a reference to that project as a dependency instead.
This does not work because of shading issues.