[maven] use parent dir for dev-tools directory
When using `dev-tools` in a sub project, maven can not find the `dev-tools` dir inside the sub module it self.

To fix that, we need to set that `elasticsearch.tools.directory` is actually using the parent dir.

```xml
<elasticsearch.tools.directory>${project.parent.build.directory}/dev-tools</elasticsearch.tools.directory>
```

Note that the issue could still appear if you are defining a sub module of the first module.

`parent->plugins->myplugin` might not be able to find it.

A workaround could be to keep `${project.build.directory}` in parent project.
But define in plugins `pom.xml` `${project.parent.build.directory}` and in myplugin `pom.xml` `${project.parent.parent.build.directory}`.